simpler start script

+++ start.sh
@@ -3,6 +3,12 @@ if (( $EUID != 1000 )); then
     echo "Must not be run as root"
     exit
 fi
-sudo ./rpm-ostree.sh
-toolbox -y create
-systemctl reboot
+if grep -q "1" "./step.log"; then
+    sudo rpm-ostree update
+    sudo rpm-ostree install ansible
+    toolbox -y create
+    echo "2" > ./step.log
+    systemctl reboot
+else
+    sudo ansible-playbook main.yml --connection=localhost -i hosts.ini
+fi
